requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
PyYAML>=6.0.2
pytest>=8.3.0
rank-bm25>=0.2.2
